Initial move ShareContentButtonConnected
