+++ packages/apollos-ui-connected/src/ShareContentButtonConnected/getShareContent.js
