+++ packages/apollos-ui-connected/src/ShareContentButtonConnected/index.js
@@ -1,0 +1,3 @@
+import ShareContentButtonConnected from './ShareContentButtonConnected';
+
+export default ShareContentButtonConnected;
